+++ src/babashka/main.clj
@@ -971,6 +971,13 @@ Use bb run --help to show this help output.
                  (and (= minor-current minor-min)
                       (>= patch-current patch-min)))))))
 
+(defn load-edn [string]
+  (try (edn/read-string string)
+    (catch java.lang.RuntimeException e
+      (if (re-find #"No dispatch macro for: \"" (.getMessage e))
+        (throw (ex-info "Invalid regex literal found in EDN config, use re-pattern instead" {}))
+        (throw e)))))
+
 (defn main [& args]
   (let [[args global-opts] (parse-global-opts args)
         {:keys [:jar] :as file-opt} (when (some-> args first io/file .isFile)
@@ -984,7 +991,7 @@ Use bb run --help to show this help output.
         bb-edn (when bb-edn-file
                  (System/setProperty "babashka.config" bb-edn-file)
                  (let [raw-string (slurp bb-edn-file)
-                       edn (edn/read-string raw-string)
+                       edn (load-edn raw-string)
                        edn (assoc edn
                                   :raw raw-string
                                   :file bb-edn-file)