+++ CHANGELOG.md
@@ -5,9 +5,11 @@ For a list of breaking changes, check [here](#breaking-changes).
 A preview of the next release can be installed from
 [babashka-dev-builds](https://github.com/babashka/babashka-dev-builds).
 
-## unreleased
+=======
+## Unreleased
 
 - Add `prepare` subcommand to download deps & pods and cache pod metadata
+- [#1041](https://github.com/babashka/babashka/issues/1041): Improve error message when regex literal in EDN config
 
 ## 0.8.0 (2022-04-04)